Make it re-runable and fix some outdated commands

+++ sources/4-backup.sh
@@ -6,6 +6,8 @@ open /Applications/Resilio\ Sync.app
 prompt
 
 pinfo "Restoring Mackup Configuration"
-ln -s ~/Documents/dotfiles/mackup-backup/.mackup.cfg ~/.mackup.cfg
-ln -s ~/Documents/dotfiles/mackup-backup/.mackup ~/.mackup
+if [ ! -f ~/.mackup.cfg ]; then
+    ln -s ~/Documents/dotfiles/mackup-backup/.mackup.cfg ~/.mackup.cfg
+    ln -s ~/Documents/dotfiles/mackup-backup/.mackup ~/.mackup
+fi
 mackup restore
